Switch to new appstream ostree branch
With the latest version of flatpak, we now default to using the new appstream branch called appstream2/$arch that has an uncompressed appstream.xml. Using the old name in the script means we never detected that flatpak pulled a new version.

Note, this means you also have to update flatpak on the backend machine.